+++ src/avalanche/test/processor_tests.cpp
@@ -101,7 +101,7 @@ struct AvalancheTestingSetup : public TestChain100Setup {
         CAddress addr(ip(GetRandInt(0xffffffff)), NODE_NONE);
         auto node =
             new CNode(id++, ServiceFlags(NODE_NETWORK), 0, INVALID_SOCKET, addr,
-                      0, 0, CAddress(), "", ConnectionType::OUTBOUND);
+                      0, 0, 0, CAddress(), "", ConnectionType::OUTBOUND);
         node->SetSendVersion(PROTOCOL_VERSION);
         node->nServices = nServices;
         m_node.peer_logic->InitializeNode(config, node);