+++ src/test/net_tests.cpp
@@ -185,14 +185,14 @@ BOOST_AUTO_TEST_CASE(cnode_simple_test) {
     CAddress addr = CAddress(CService(ipv4Addr, 7777), NODE_NETWORK);
     std::string pszDest;
 
-    auto pnode1 =
-        std::make_unique<CNode>(id++, NODE_NETWORK, height, hSocket, addr, 0, 0,
-                                CAddress(), pszDest, ConnectionType::OUTBOUND);
+    auto pnode1 = std::make_unique<CNode>(id++, NODE_NETWORK, height, hSocket,
+                                          addr, 0, 0, 0, CAddress(), pszDest,
+                                          ConnectionType::OUTBOUND);
     BOOST_CHECK(pnode1->IsInboundConn() == false);
 
-    auto pnode2 =
-        std::make_unique<CNode>(id++, NODE_NETWORK, height, hSocket, addr, 1, 1,
-                                CAddress(), pszDest, ConnectionType::INBOUND);
+    auto pnode2 = std::make_unique<CNode>(id++, NODE_NETWORK, height, hSocket,
+                                          addr, 1, 1, 1, CAddress(), pszDest,
+                                          ConnectionType::INBOUND);
     BOOST_CHECK(pnode2->IsInboundConn() == true);
 }
 
@@ -316,7 +316,7 @@ BOOST_AUTO_TEST_CASE(ipv4_peer_with_ipv6_addrMe_test) {
     ipv4AddrPeer.s_addr = 0xa0b0c001;
     CAddress addr = CAddress(CService(ipv4AddrPeer, 7777), NODE_NETWORK);
     std::unique_ptr<CNode> pnode = std::make_unique<CNode>(
-        0, NODE_NETWORK, 0, INVALID_SOCKET, addr, 0, 0, CAddress{},
+        0, NODE_NETWORK, 0, INVALID_SOCKET, addr, 0, 0, 0, CAddress{},
         std::string{}, ConnectionType::OUTBOUND);
     pnode->fSuccessfullyConnected.store(true);