+++ doc/release-notes.md
@@ -5,3 +5,4 @@ Bitcoin ABC version 0.22.12 is now available from:
   <https://download.bitcoinabc.org/0.22.12/>
 
 This release includes the following features and fixes:
+ - Add an extra 64 bits of entropy in the initial version message.